Implementando service e chamada http.

+++ src/app/pages/register/register.component.ts
@@ -2,6 +2,9 @@
 import { Component, OnInit } from '@angular/core';
 import Swal from 'sweetalert2';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import { msg } from 'src/app/shared/utils/msg';
+import { RegisterService } from './services/register.service';
+import { IUser } from './interfaces/IUser';
 
 @Component({
   selector: 'app-register',
@@ -10,8 +13,8 @@ import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 })
 export class RegisterComponent implements OnInit {
 
-  constructor(private fb: FormBuilder) { }
-
+  constructor(private fb: FormBuilder, private registerService: RegisterService) { }
+  msg = msg;
   registerForm: FormGroup = this.fb.group({
     role: ['', [Validators.required]],
     fullName: ['', [Validators.required]],
@@ -36,53 +39,42 @@ export class RegisterComponent implements OnInit {
     return counter !== list.length;
   }
 
+  toogleRole(role: 'dev'|'cliente') {
+    this.registerForm.get('role')?.setValue(role)
+  }
+
   cadastrar() {
-    //Verifica se alguma role está checada
-    if(this.roleIsChecked() === false) {
-      Swal.fire(
-        'Algo deu errado...',
-        'Marque alguma role!',
-        'error'
-      )
-      return;
+    if (this.registerForm.valid) {
+      let payload: IUser = this.registerForm.value;
+      this.registerService.postUser(payload).subscribe({
+        next: (response) => {
+          Swal.fire({
+          title: 'Bom trabalho!',
+          text: 'Cadastrado com sucesso!',
+          icon: 'success',
+          confirmButtonText: 'OK!'
+        }).then((result) => {
+          if(result.isConfirmed) {
+            localStorage.setItem('user.name', response.fullName);
+            localStorage.setItem('role', response.role === 'Dev' ? 'Desenvolvedor' : 'Cliente');
+            localStorage.setItem('clientId', response.id);
+            window.location.href = "list.html";
+          }
+        })
+      },
+      error: (error) => {
+          console.log(error);
+        },
+      });
+    } else {
+      this.registerForm.markAllAsTouched();
     }
+  }
 
-    //Pegar os dados do formulário
-    let payLoad = {
-      role: document.getElementsByName("cargo")[0].checked === true ? 'Dev' : 'Cliente',
-      fullName: document.querySelector("#fullName").value,
-      birthDate: document.querySelector("#birthDate").value,
-      email: document.querySelector("#email").value,
-      password: document.querySelector("#password").value
+  isInvalid(inputName: string, validatorName: string) {
+    const formControl: any = this.registerForm.get(inputName);
+    if(formControl.errors !== null) {
+      return formControl.errors[validatorName] && formControl.touched;
     }
-
-    //Enviar para a api
-    fetch("https://637c0ed76f4024eac21d48b8.mockapi.io/api/users", {
-      method: 'POST',
-      body: JSON.stringify(payLoad),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    })
-    .then(response => response.json())
-    .then(response => {
-      Swal.fire({
-        title: 'Bom trabalho!',
-        text: 'Cadastrado com sucesso!',
-        icon: 'success',
-        confirmButtonText: 'OK!'
-      }).then((result) => {
-        if(result.isConfirmed) {
-          localStorage.setItem('user.name', response.fullName);
-          localStorage.setItem('role', response.role === 'Dev' ? 'Desenvolvedor' : 'Cliente');
-          localStorage.setItem('clientId', response.id);
-          window.location.href = "list.html";
-        }
-      })
-    })
-    .catch(error => {
-      alert('Erro no servidor!');
-      console.log(error);
-    })
   }
 }
